Provide option to ignore missing values
This library is awesome and a cornerstone of all our data access. However, one thing we've continually found ourselves wishing for is the ability to ignore missing params and supply PG with a null value. 

This is a crud example, written in the Github editor, but feels like it'd cover most of what we'd like to see. We don't use any of the file scanning, so would love some feedback and direction on how this could be written to make it into core. Happy to write tests / docs as needed. 

Thanks for the great OSS contribution :)